Ecl:Ptzp: Pattern class command update
Data getting from user on pattern API was an error.
if the data type is float or double type , this was a mistake.
Because integer converter used. This problem solved.
Now, Float converters and variables are using for pattern

+++ patternng.cpp
@@ -103,7 +103,7 @@ void PatternNg::positionUpdate(int pan, int tilt, int zoom)
 	}
 }
 
-void PatternNg::commandUpdate(int pan, int tilt, int zoom, int c, int par1, int par2)
+void PatternNg::commandUpdate(int pan, int tilt, int zoom, int c, float par1, float par2)
 {
 	if (!isRecording()) {
 		return;
